fix: :bug: delete render counting
