+++ src/utils/todoItem.js
@@ -20,6 +20,8 @@ function findTodoItem(todoItems, id) {
 function findTodoItemIndex(todoItems, id) {
   let targetIndex = -1;
 
+  console.log(todoItems);
+
   todoItems.forEach((todoItem, index) => {
     if (todoItem.id === id) {
       targetIndex = index;
